requirements.txt
firebase-admin~=6.6.0
requests
dhanhq~=2.0.2
python-dotenv~=1.0.1
pandas~=2.2.3
numpy~=2.2.2
ta~=0.11.0
kiteconnect~=5.0.1
scikit-learn~=1.7.0